requirements.txt
feedparser>=6.0.10
pyyaml>=6.0.1
requests>=2.31.0
python-dateutil>=2.8.2
openai>=1.35.0
